Make custom hook for notification
