+++ frontend/src/Components/SideBar.jsx
@@ -9,10 +9,7 @@ import { IoIosHelpCircleOutline } from "react-icons/io";
 import { BiUserCircle } from "react-icons/bi";
 import { CiLogout } from "react-icons/ci";
 import authService from "../services/authService";
-import { toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
-import { setShowToast } from "../reducers/toastSlice";
-import { useDispatch, useSelector } from "react-redux";
+import useToast from "../hooks/useToast";
 
 const SidebarContainer = styled.div`
   background-color: #ffffff;
@@ -86,16 +83,7 @@ const LogoutButton = styled.button`
 
 const SideBar = () => {
   const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const showToast = useSelector((state) => state.toast.showToast);
-
-  const notify = (message) => {
-    if (showToast) return;
-    toast(message, {
-      onOpen: () => dispatch(setShowToast(true)),
-      onClose: () => dispatch(setShowToast(false)),
-    });
-  };
+  const { notify } = useToast();
 
   const handleLogout = async (e) => {
     e.preventDefault();
