+++ frontend/src/services/storageService.js
@@ -1,0 +1,2 @@
+//TODO. Add here all the methods that you need to access the storage service, i.e. reading and writing to firebase storage.
+//TODO. You can use the methods that are already implemented in the authService.js file as an example.
